+++ Addon/ElvUI_mMediaTag/options/changelog/3.48.lua
@@ -1,9 +1,0 @@
-local mMT, DB, M, E, P, L, MEDIA = unpack(ElvUI_mMediaTag)
-
-mMT.Changelog[348] = {
-	DATE = "07.08.2024",
-	FIX = {
-		"[Datatext]: Bug with Game Menu Datatext and Customer Support window auto opens.",
-        "[System]: Dropdown Menus on Classic, API changes.",
-	},
-}