+++ Addon/ElvUI_mMediaTag/options/changelog/3.58.lua
@@ -1,9 +1,0 @@
-local mMT, DB, M, E, P, L, MEDIA = unpack(ElvUI_mMediaTag)
-
-mMT.Changelog[358]  = {
-	DATE = "11.09.2024",
-	NEW = {
-		"[Portraits]: Add a function to add your custom Class Icons to the portraits via code.",
-		"[Portraits]: Add e Set of new Class Icons.",
-	},
-}