+++ Addon/ElvUI_mMediaTag/options/changelog/3.62.lua
@@ -1,9 +1,0 @@
-local mMT, DB, M, E, P, L, MEDIA = unpack(ElvUI_mMediaTag)
-
-mMT.Changelog[362]  = {
-	DATE = "14.10.2024",
-	FIX = {
-		"[Objectivtracker]: Skin for Delves and Dungeons.",
-		"[Objectivtracker]: Removed Test functions.",
-	},
-}