+++ Addon/ElvUI_mMediaTag/options/changelog/4.00.lua
@@ -1,0 +1,20 @@
+local mMT, DB, M, E, P, L, MEDIA = unpack(ElvUI_mMediaTag)
+
+mMT.Changelog[400] = {
+	DATE = "13.03.2026",
+	IMPORTANT = {
+		"[INFO]: A completely new update for Midnight, almost everything has been optimized and restructured.",
+		"[INFO]: Some modules and features are still missing, but this is a fully functional and working version for Midnight. There will be no version for Classic, where the 3.xx version should still work. I will continue to add more features over time.",
+	},
+	UPDATE = {
+		"[Portraits]: Many new Textures and removed Old versions.",
+		"[Intrrupt-On-CD]: Update for Midnight.",
+		"[Unitframe-Icons]: New Icons and removed some old ones.",
+		"[Nameplate]: Auto set Target glow color is Back.",
+		"[Nameplate]: Highlighter for Quest, Focus and Target Units.",
+		"[Datatext]: Reworked almost all Datatext, you can Now easily Track custom items or currency's.",
+		"[Datatext]: IDs and Teleports for Midnight.",
+		"[Dock]: Update Textures and removed some.",
+		"[TAGs]: Update for Midnight and removed the ones that can't be done now in Midnight. Cleanup.",
+	},
+}