+++ Addon/ElvUI_mMediaTag/profiles/elvui_defaults.lua
@@ -2,6 +2,7 @@ local mMT, DB, M, E, P, L, MEDIA = unpack(ElvUI_mMediaTag)
 
 -- system settings
 P.db_version = nil
+P.version = nil
 
 -- general
 P.general = {