+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 
 ## [ver. 4.00 - 13.03.2026
 ### UPDATE
-- UPDATE - [System]: A completely new update for Midnight, almost everything has been optimized and restructured.
+- UPDATE - [Portraits]: Many new Textures and removed Old versions.
+- UPDATE - [Intrrupt-On-CD]: Update for Midnight.
+- UPDATE - [Unitframe-Icons]: New Icons and removed some old ones.
+- UPDATE - [Nameplate]: Auto set Target glow color is Back.
+- UPDATE - [Nameplate]: Highlighter for Quest, Focus and Target Units.
+- UPDATE - [Datatext]: Reworked almost all Datatext, you can Now easily Track custom items or currency's.
+- UPDATE - [Datatext]: IDs and Teleports for Midnight.
+- UPDATE - [Dock]: Update Textures and removed some.
+- UPDATE - [TAGs]: Update for Midnight and removed the ones that can't be done now in Midnight. Cleanup.
 ### INFO
+- INFO - A completely new update for Midnight, almost everything has been optimized and restructured.
 - INFO - Some modules and features are still missing, but this is a fully functional and working version for Midnight. There will be no version for Classic, where the 3.xx version should still work. I will continue to add more features over time.